ExperimentJob {BatchExperiments}	R Documentation

ExperimentJob

Description

You can access job properties using the job object which is optionally passed to dynamic problem functions and algorithms. The object is a named list with the following elements:

id [integer(1)]:

Job ID.

prob.id [character(1)]:

Problem ID.

prob.pars [list]:

Problem parameters as named list.

algo.id [character(1)]:

algo.id

Algorithm ID.

algo.pars [list]:

Algorithm parameters as named list.

repl [integer(1)]:

Replication number of this experiment.

seed [integer(1)]:

Seed set right before algorithm execution.

prob.seed [integer(1)]:

Seed set right before generation of problem instance.

[Package BatchExperiments version 1.1 Index]