addAlgorithm {BatchExperiments}	R Documentation

Add an algorithm to registry.

Description

Add an algorithm to registry and stores it on disk.

Usage

addAlgorithm(reg, id, fun, overwrite = FALSE)

Arguments

reg
[ExperimentRegistry]
Registry.

id
[character(1)]
Name of algorithm.

fun
[function(job, static, dynamic, ...)]
Function which applies the algorithm to a problem instance. Takes a Job object, the static problem part and the evaluated dynamic problem part as arguments. You may omit any of job, static or dynamic. In this case, the respective arguments will not get passed to fun. Further parameters from Design are passed to ... argument. If you are using multiple result files this function must return a named list.

To retrieve job informations from the job object see the documentation on ExperimentJob.

overwrite
[logical(1)]
Overwrite the algorithm file if it already exists? Default is FALSE.

Value

[character(1)]. Invisibly returns the id.

[Package BatchExperiments version 1.1 Index]