addExperiments {BatchExperiments}	R Documentation

Add experiemts to the registry.

Description

Add experiments for running algorithms on problems to the registry, so they can be executed later.

Usage

addExperiments(reg, prob.designs, algo.designs, repls = 1L,
  skip.defined = FALSE)

Arguments

reg
[ExperimentRegistry]
Registry.

prob.designs
[character | Design | list of Design]
Either problem ids, a single problem design or a list of problem designs, the latter two created by makeDesign. If missing, all problems are selected (without associating a design), and this is the default.

algo.designs
[character | Design | list of Design]
Either algorithm ids, a single algorithm design or a list of algorithm designs, the latter two created by makeDesign. If missing, all algorithms are selected (without associating a design), and this is the default.

repls
[integer(1)]
Number of replications.
Default is 1.

skip.defined
[logical]
If set to TRUE, already defined experiments get skipped. Otherwise an error is thrown.
Default is FALSE.

Value

Invisibly returns vector of ids of added experiments.

Examples

### EXAMPLE 1 ###
reg <- makeExperimentRegistry(id="example1", file.dir=tempfile())

## Creating dir: /tmp/RtmpVeetNu/file576b6e07083a
## Saving registry: /tmp/RtmpVeetNu/file576b6e07083a/registry.RData

# Define a problem:
# Subsampling from the iris dataset.
data(iris)
subsample <- function(static, ratio) {
  n <- nrow(static)
  train <- sample(n, floor(n * ratio))
  test <- setdiff(seq(n), train)
  list(test=test, train=train)
}
addProblem(reg, id="iris", static=iris,
           dynamic=subsample, seed=123)

## Writing problem files: /tmp/RtmpVeetNu/file576b6e07083a/problems/iris_static.RData, /tmp/RtmpVeetNu/file576b6e07083a/problems/iris_dynamic.RData

# Define algorithm "tree":
# Decision tree on the iris dataset, modeling Species.
tree.wrapper <- function(static, dynamic, ...) {
  library(rpart)
  mod <- rpart(Species ~ ., data=static[dynamic$train, ], ...)
  pred <- predict(mod, newdata=static[dynamic$test, ], type="class")
  table(static$Species[dynamic$test], pred)
}
addAlgorithm(reg, id="tree", fun=tree.wrapper)

## Writing algorithm file: /tmp/RtmpVeetNu/file576b6e07083a/algorithms/tree.RData

# Define algorithm "forest":
# Random forest on the iris dataset, modeling Species.
forest.wrapper <- function(static, dynamic, ...) {
  library(randomForest)
  mod <- randomForest(Species ~ ., data=static, subset=dynamic$train, ...)
  pred <- predict(mod, newdata=static[dynamic$test, ])
  table(static$Species[dynamic$test], pred)
}
addAlgorithm(reg, id="forest", fun=forest.wrapper)

## Writing algorithm file: /tmp/RtmpVeetNu/file576b6e07083a/algorithms/forest.RData

# Define problem parameters:
pars <- list(ratio=c(0.67, 0.9))
iris.design <- makeDesign("iris", exhaustive=pars)

# Define decision tree parameters:
pars <- list(minsplit=c(10, 20), cp=c(0.01, 0.1))
tree.design <- makeDesign("tree", exhaustive=pars)

# Define random forest parameters:
pars <- list(ntree=c(100, 500))
forest.design <- makeDesign("forest", exhaustive=pars)

# Add experiments to the registry:
# Use  previously defined experimental designs.
addExperiments(reg, prob.designs=iris.design,
               algo.designs=list(tree.design, forest.design),
               repls=2) # usually you would set repls to 100 or more.

## Adding 12 experiments / 24 jobs to DB.

# Optional: Short summary over problems and algorithms.
summarizeExperiments(reg)

##   prob   algo .count
## 1 iris forest      8
## 2 iris   tree     16

# Optional: Test one decision tree job and one expensive (ntree=1000)
# random forest job. Use findExperiments to get the right job ids.
do.tests <- FALSE
if (do.tests) {
  id1 <- findExperiments(reg, algo.pattern="tree")[1]
  id2 <- findExperiments(reg, algo.pattern="forest",
                         algo.pars=(ntree == 1000))[1]
  testJob(reg, id1)
  testJob(reg, id2)
}

# Submit the jobs to the batch system
submitJobs(reg)

## Saving conf: /tmp/RtmpVeetNu/file576b6e07083a/conf.RData
## Submitting 24 chunks / 24 jobs.
## Cluster functions: Interactive.
## Auto-mailer settings: start=none, done=none, error=none.
## Writing 24 R scripts...
## 2014-02-07 19:02:16: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/file576b6e07083a/registry.RData
## Loading conf: /tmp/RtmpVeetNu/file576b6e07083a/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=1 ##########
## Timestamp: 2014-02-07 19:02:16
## Setting seed: 700935866
## Applying Algorithm tree ...
## Generating problem iris ...
## Writing result file: /tmp/RtmpVeetNu/file576b6e07083a/jobs/01/1-result.RData
## 2014-02-07 19:02:16: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:16: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/file576b6e07083a/registry.RData
## Loading conf: /tmp/RtmpVeetNu/file576b6e07083a/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=2 ##########
## Timestamp: 2014-02-07 19:02:16
## Setting seed: 700935867
## Applying Algorithm tree ...
## Generating problem iris ...
## Writing result file: /tmp/RtmpVeetNu/file576b6e07083a/jobs/02/2-result.RData
## 2014-02-07 19:02:16: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:17: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/file576b6e07083a/registry.RData
## Loading conf: /tmp/RtmpVeetNu/file576b6e07083a/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=3 ##########
## Timestamp: 2014-02-07 19:02:17
## Setting seed: 700935868
## Applying Algorithm tree ...
## Generating problem iris ...
## Writing result file: /tmp/RtmpVeetNu/file576b6e07083a/jobs/03/3-result.RData
## 2014-02-07 19:02:17: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:17: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/file576b6e07083a/registry.RData
## Loading conf: /tmp/RtmpVeetNu/file576b6e07083a/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=4 ##########
## Timestamp: 2014-02-07 19:02:17
## Setting seed: 700935869
## Applying Algorithm tree ...
## Generating problem iris ...
## Writing result file: /tmp/RtmpVeetNu/file576b6e07083a/jobs/04/4-result.RData
## 2014-02-07 19:02:17: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:17: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/file576b6e07083a/registry.RData
## Loading conf: /tmp/RtmpVeetNu/file576b6e07083a/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=5 ##########
## Timestamp: 2014-02-07 19:02:17
## Setting seed: 700935870
## Applying Algorithm tree ...
## Generating problem iris ...
## Writing result file: /tmp/RtmpVeetNu/file576b6e07083a/jobs/05/5-result.RData
## 2014-02-07 19:02:17: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:17: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/file576b6e07083a/registry.RData
## Loading conf: /tmp/RtmpVeetNu/file576b6e07083a/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=6 ##########
## Timestamp: 2014-02-07 19:02:17
## Setting seed: 700935871
## Applying Algorithm tree ...
## Generating problem iris ...
## Writing result file: /tmp/RtmpVeetNu/file576b6e07083a/jobs/06/6-result.RData
## 2014-02-07 19:02:17: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:17: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/file576b6e07083a/registry.RData
## Loading conf: /tmp/RtmpVeetNu/file576b6e07083a/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=7 ##########
## Timestamp: 2014-02-07 19:02:17
## Setting seed: 700935872
## Applying Algorithm tree ...
## Generating problem iris ...
## Writing result file: /tmp/RtmpVeetNu/file576b6e07083a/jobs/07/7-result.RData
## 2014-02-07 19:02:17: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:18: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/file576b6e07083a/registry.RData
## Loading conf: /tmp/RtmpVeetNu/file576b6e07083a/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=8 ##########
## Timestamp: 2014-02-07 19:02:18
## Setting seed: 700935873
## Applying Algorithm tree ...
## Generating problem iris ...
## Writing result file: /tmp/RtmpVeetNu/file576b6e07083a/jobs/08/8-result.RData
## 2014-02-07 19:02:18: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:18: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/file576b6e07083a/registry.RData
## Loading conf: /tmp/RtmpVeetNu/file576b6e07083a/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=9 ##########
## Timestamp: 2014-02-07 19:02:18
## Setting seed: 700935874
## Applying Algorithm forest ...
## randomForest 4.6-7
## Type rfNews() to see new features/changes/bug fixes.
## Generating problem iris ...
## Writing result file: /tmp/RtmpVeetNu/file576b6e07083a/jobs/09/9-result.RData
## 2014-02-07 19:02:18: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:18: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/file576b6e07083a/registry.RData
## Loading conf: /tmp/RtmpVeetNu/file576b6e07083a/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=10 ##########
## Timestamp: 2014-02-07 19:02:18
## Setting seed: 700935875
## Applying Algorithm forest ...
## Generating problem iris ...
## Writing result file: /tmp/RtmpVeetNu/file576b6e07083a/jobs/10/10-result.RData
## 2014-02-07 19:02:18: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:18: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/file576b6e07083a/registry.RData
## Loading conf: /tmp/RtmpVeetNu/file576b6e07083a/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=11 ##########
## Timestamp: 2014-02-07 19:02:18
## Setting seed: 700935876
## Applying Algorithm forest ...
## Generating problem iris ...
## Writing result file: /tmp/RtmpVeetNu/file576b6e07083a/jobs/11/11-result.RData
## 2014-02-07 19:02:19: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:19: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/file576b6e07083a/registry.RData
## Loading conf: /tmp/RtmpVeetNu/file576b6e07083a/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=12 ##########
## Timestamp: 2014-02-07 19:02:19
## Setting seed: 700935877
## Applying Algorithm forest ...
## Generating problem iris ...
## Writing result file: /tmp/RtmpVeetNu/file576b6e07083a/jobs/12/12-result.RData
## 2014-02-07 19:02:19: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:19: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/file576b6e07083a/registry.RData
## Loading conf: /tmp/RtmpVeetNu/file576b6e07083a/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=13 ##########
## Timestamp: 2014-02-07 19:02:19
## Setting seed: 700935878
## Applying Algorithm tree ...
## Generating problem iris ...
## Writing result file: /tmp/RtmpVeetNu/file576b6e07083a/jobs/13/13-result.RData
## 2014-02-07 19:02:19: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:19: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/file576b6e07083a/registry.RData
## Loading conf: /tmp/RtmpVeetNu/file576b6e07083a/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=14 ##########
## Timestamp: 2014-02-07 19:02:19
## Setting seed: 700935879
## Applying Algorithm tree ...
## Generating problem iris ...
## Writing result file: /tmp/RtmpVeetNu/file576b6e07083a/jobs/14/14-result.RData
## 2014-02-07 19:02:19: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:19: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/file576b6e07083a/registry.RData
## Loading conf: /tmp/RtmpVeetNu/file576b6e07083a/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=15 ##########
## Timestamp: 2014-02-07 19:02:19
## Setting seed: 700935880
## Applying Algorithm tree ...
## Generating problem iris ...
## Writing result file: /tmp/RtmpVeetNu/file576b6e07083a/jobs/15/15-result.RData
## 2014-02-07 19:02:20: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:20: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/file576b6e07083a/registry.RData
## Loading conf: /tmp/RtmpVeetNu/file576b6e07083a/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=16 ##########
## Timestamp: 2014-02-07 19:02:20
## Setting seed: 700935881
## Applying Algorithm tree ...
## Generating problem iris ...
## Writing result file: /tmp/RtmpVeetNu/file576b6e07083a/jobs/16/16-result.RData
## 2014-02-07 19:02:20: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:20: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/file576b6e07083a/registry.RData
## Loading conf: /tmp/RtmpVeetNu/file576b6e07083a/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=17 ##########
## Timestamp: 2014-02-07 19:02:20
## Setting seed: 700935882
## Applying Algorithm tree ...
## Generating problem iris ...
## Writing result file: /tmp/RtmpVeetNu/file576b6e07083a/jobs/17/17-result.RData
## 2014-02-07 19:02:20: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:20: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/file576b6e07083a/registry.RData
## Loading conf: /tmp/RtmpVeetNu/file576b6e07083a/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=18 ##########
## Timestamp: 2014-02-07 19:02:20
## Setting seed: 700935883
## Applying Algorithm tree ...
## Generating problem iris ...
## Writing result file: /tmp/RtmpVeetNu/file576b6e07083a/jobs/18/18-result.RData
## 2014-02-07 19:02:20: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:20: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/file576b6e07083a/registry.RData
## Loading conf: /tmp/RtmpVeetNu/file576b6e07083a/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=19 ##########
## Timestamp: 2014-02-07 19:02:20
## Setting seed: 700935884
## Applying Algorithm tree ...
## Generating problem iris ...
## Writing result file: /tmp/RtmpVeetNu/file576b6e07083a/jobs/19/19-result.RData
## 2014-02-07 19:02:20: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:20: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/file576b6e07083a/registry.RData
## Loading conf: /tmp/RtmpVeetNu/file576b6e07083a/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=20 ##########
## Timestamp: 2014-02-07 19:02:21
## Setting seed: 700935885
## Applying Algorithm tree ...
## Generating problem iris ...
## Writing result file: /tmp/RtmpVeetNu/file576b6e07083a/jobs/20/20-result.RData
## 2014-02-07 19:02:21: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:21: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/file576b6e07083a/registry.RData
## Loading conf: /tmp/RtmpVeetNu/file576b6e07083a/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=21 ##########
## Timestamp: 2014-02-07 19:02:21
## Setting seed: 700935886
## Applying Algorithm forest ...
## Generating problem iris ...
## Writing result file: /tmp/RtmpVeetNu/file576b6e07083a/jobs/21/21-result.RData
## 2014-02-07 19:02:21: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:21: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/file576b6e07083a/registry.RData
## Loading conf: /tmp/RtmpVeetNu/file576b6e07083a/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=22 ##########
## Timestamp: 2014-02-07 19:02:21
## Setting seed: 700935887
## Applying Algorithm forest ...
## Generating problem iris ...
## Writing result file: /tmp/RtmpVeetNu/file576b6e07083a/jobs/22/22-result.RData
## 2014-02-07 19:02:21: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:21: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/file576b6e07083a/registry.RData
## Loading conf: /tmp/RtmpVeetNu/file576b6e07083a/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=23 ##########
## Timestamp: 2014-02-07 19:02:21
## Setting seed: 700935888
## Applying Algorithm forest ...
## Generating problem iris ...
## Writing result file: /tmp/RtmpVeetNu/file576b6e07083a/jobs/23/23-result.RData
## 2014-02-07 19:02:21: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:22: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/file576b6e07083a/registry.RData
## Loading conf: /tmp/RtmpVeetNu/file576b6e07083a/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=24 ##########
## Timestamp: 2014-02-07 19:02:22
## Setting seed: 700935889
## Applying Algorithm forest ...
## Generating problem iris ...
## Writing result file: /tmp/RtmpVeetNu/file576b6e07083a/jobs/24/24-result.RData
## 2014-02-07 19:02:22: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:

## 
SubmitJobs |+                                                |   0% (00:00:00)
SubmitJobs |++                                               |   4% (00:00:00)
SubmitJobs |++++                                             |   8% (00:00:11)
SubmitJobs |++++++                                           |  12% (00:00:07)
SubmitJobs |++++++++                                         |  17% (00:00:05)
SubmitJobs |++++++++++                                       |  21% (00:00:03)
SubmitJobs |++++++++++++                                     |  25% (00:00:03)
SubmitJobs |++++++++++++++                                   |  29% (00:00:04)
SubmitJobs |++++++++++++++++                                 |  33% (00:00:04)
SubmitJobs |++++++++++++++++++                               |  38% (00:00:03)
SubmitJobs |++++++++++++++++++++                             |  42% (00:00:02)
SubmitJobs |++++++++++++++++++++++                           |  46% (00:00:03)
SubmitJobs |++++++++++++++++++++++++                         |  50% (00:00:03)
SubmitJobs |+++++++++++++++++++++++++++                      |  54% (00:00:02)
SubmitJobs |+++++++++++++++++++++++++++++                    |  58% (00:00:02)
SubmitJobs |+++++++++++++++++++++++++++++++                  |  62% (00:00:02)
SubmitJobs |+++++++++++++++++++++++++++++++++                |  67% (00:00:02)
SubmitJobs |+++++++++++++++++++++++++++++++++++              |  71% (00:00:01)
SubmitJobs |+++++++++++++++++++++++++++++++++++++            |  75% (00:00:01)
SubmitJobs |+++++++++++++++++++++++++++++++++++++++          |  79% (00:00:01)
SubmitJobs |+++++++++++++++++++++++++++++++++++++++++        |  83% (00:00:01)
SubmitJobs |+++++++++++++++++++++++++++++++++++++++++++      |  88% (00:00:00)
SubmitJobs |+++++++++++++++++++++++++++++++++++++++++++++    |  92% (00:00:00)
SubmitJobs |+++++++++++++++++++++++++++++++++++++++++++++++  |  96% (00:00:00)
SubmitJobs |+++++++++++++++++++++++++++++++++++++++++++++++++| 100% (00:00:00)

## Sending 24 submit messages...
## Might take some time, do not interrupt this!

# Calculate the misclassification rate for all (already done) jobs.
reduce <- function(job, res) {
  n <- sum(res)
  list(mcr=(n-sum(diag(res)))/n)
}
res <- reduceResultsExperiments(reg, fun=reduce)

## Reducing 24 results...

## 
reduceResultsExperiments |+                                  |   0% (00:00:00)
reduceResultsExperiments |+++++++++++++++++++++++++++++++++++| 100% (00:00:00)

print(res)

##    id prob ratio   algo   cp minsplit repl     mcr ntree
## 1   1 iris  0.67   tree 0.01       10    1 0.08000    NA
## 2   2 iris  0.67   tree 0.01       10    2 0.06000    NA
## 3   3 iris  0.67   tree 0.01       20    1 0.08000    NA
## 4   4 iris  0.67   tree 0.01       20    2 0.06000    NA
## 5   5 iris  0.67   tree 0.10       10    1 0.08000    NA
## 6   6 iris  0.67   tree 0.10       10    2 0.06000    NA
## 7   7 iris  0.67   tree 0.10       20    1 0.08000    NA
## 8   8 iris  0.67   tree 0.10       20    2 0.06000    NA
## 9   9 iris  0.67 forest   NA       NA    1 0.06000   100
## 10 10 iris  0.67 forest   NA       NA    2 0.04000   100
## 11 11 iris  0.67 forest   NA       NA    1 0.06000   500
## 12 12 iris  0.67 forest   NA       NA    2 0.06000   500
## 13 13 iris  0.90   tree 0.01       10    1 0.00000    NA
## 14 14 iris  0.90   tree 0.01       10    2 0.06667    NA
## 15 15 iris  0.90   tree 0.01       20    1 0.00000    NA
## 16 16 iris  0.90   tree 0.01       20    2 0.06667    NA
## 17 17 iris  0.90   tree 0.10       10    1 0.00000    NA
## 18 18 iris  0.90   tree 0.10       10    2 0.00000    NA
## 19 19 iris  0.90   tree 0.10       20    1 0.00000    NA
## 20 20 iris  0.90   tree 0.10       20    2 0.00000    NA
## 21 21 iris  0.90 forest   NA       NA    1 0.00000   100
## 22 22 iris  0.90 forest   NA       NA    2 0.06667   100
## 23 23 iris  0.90 forest   NA       NA    1 0.00000   500
## 24 24 iris  0.90 forest   NA       NA    2 0.06667   500

# Aggregate results using 'ddply' from package 'plyr':
# Calculate the mean over all replications of identical experiments
# (same problem, same algorithm and same parameters)
library(plyr)
vars <- setdiff(names(res), c("repl", "mcr"))
aggr <- ddply(res, vars, summarise, mean.mcr=mean(mcr))
print(aggr)

##    id prob ratio   algo   cp minsplit ntree mean.mcr
## 1   1 iris  0.67   tree 0.01       10    NA  0.08000
## 2   2 iris  0.67   tree 0.01       10    NA  0.06000
## 3   3 iris  0.67   tree 0.01       20    NA  0.08000
## 4   4 iris  0.67   tree 0.01       20    NA  0.06000
## 5   5 iris  0.67   tree 0.10       10    NA  0.08000
## 6   6 iris  0.67   tree 0.10       10    NA  0.06000
## 7   7 iris  0.67   tree 0.10       20    NA  0.08000
## 8   8 iris  0.67   tree 0.10       20    NA  0.06000
## 9   9 iris  0.67 forest   NA       NA   100  0.06000
## 10 10 iris  0.67 forest   NA       NA   100  0.04000
## 11 11 iris  0.67 forest   NA       NA   500  0.06000
## 12 12 iris  0.67 forest   NA       NA   500  0.06000
## 13 13 iris  0.90   tree 0.01       10    NA  0.00000
## 14 14 iris  0.90   tree 0.01       10    NA  0.06667
## 15 15 iris  0.90   tree 0.01       20    NA  0.00000
## 16 16 iris  0.90   tree 0.01       20    NA  0.06667
## 17 17 iris  0.90   tree 0.10       10    NA  0.00000
## 18 18 iris  0.90   tree 0.10       10    NA  0.00000
## 19 19 iris  0.90   tree 0.10       20    NA  0.00000
## 20 20 iris  0.90   tree 0.10       20    NA  0.00000
## 21 21 iris  0.90 forest   NA       NA   100  0.00000
## 22 22 iris  0.90 forest   NA       NA   100  0.06667
## 23 23 iris  0.90 forest   NA       NA   500  0.00000
## 24 24 iris  0.90 forest   NA       NA   500  0.06667

### EXAMPLE 2 ###
# define two simple test functions
testfun1 = function(x) sum(x^2)
testfun2 = function(x) -exp(-sum(abs(x)))

# Define ExperimentRegistry:
reg = makeExperimentRegistry("example02", seed=123, file.dir=tempfile())

## Creating dir: /tmp/RtmpVeetNu/file576b3d599d5c
## Saving registry: /tmp/RtmpVeetNu/file576b3d599d5c/registry.RData

# Add the testfunctions to the registry:
addProblem(reg, "testfun1", static=testfun1)

## Writing problem files: /tmp/RtmpVeetNu/file576b3d599d5c/problems/testfun1_static.RData, /tmp/RtmpVeetNu/file576b3d599d5c/problems/testfun1_dynamic.RData

addProblem(reg, "testfun2", static=testfun2)

## Writing problem files: /tmp/RtmpVeetNu/file576b3d599d5c/problems/testfun2_static.RData, /tmp/RtmpVeetNu/file576b3d599d5c/problems/testfun2_dynamic.RData

# Use SimulatedAnnealing on the test functions:
addAlgorithm(reg, "sann", fun=function(static, dynamic) {
  upp = rep(10, 2)
  low = -upp
  start = sample(c(-10, 10), 2)
  res = optim(start, fn=static, lower=low, upper=upp, method="SANN")
  res = res[c("par", "value", "counts", "convergence")]
  res$start = start
  return(res)
})

## Writing algorithm file: /tmp/RtmpVeetNu/file576b3d599d5c/algorithms/sann.RData

# add experiments and submit
addExperiments(reg, repls=10)

## Adding 2 experiments / 20 jobs to DB.

submitJobs(reg)

## Saving conf: /tmp/RtmpVeetNu/file576b3d599d5c/conf.RData
## Submitting 20 chunks / 20 jobs.
## Cluster functions: Interactive.
## Auto-mailer settings: start=none, done=none, error=none.
## Writing 20 R scripts...
## 2014-02-07 19:02:23: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/file576b3d599d5c/registry.RData
## Loading conf: /tmp/RtmpVeetNu/file576b3d599d5c/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=1 ##########
## Timestamp: 2014-02-07 19:02:23
## Setting seed: 123
## Applying Algorithm sann ...

## Warning: bounds can only be used with method L-BFGS-B (or Brent)

## Writing result file: /tmp/RtmpVeetNu/file576b3d599d5c/jobs/01/1-result.RData
## 2014-02-07 19:02:23: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:23: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/file576b3d599d5c/registry.RData
## Loading conf: /tmp/RtmpVeetNu/file576b3d599d5c/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=2 ##########
## Timestamp: 2014-02-07 19:02:23
## Setting seed: 124
## Applying Algorithm sann ...

## Warning: bounds can only be used with method L-BFGS-B (or Brent)

## Writing result file: /tmp/RtmpVeetNu/file576b3d599d5c/jobs/02/2-result.RData
## 2014-02-07 19:02:23: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:23: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/file576b3d599d5c/registry.RData
## Loading conf: /tmp/RtmpVeetNu/file576b3d599d5c/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=3 ##########
## Timestamp: 2014-02-07 19:02:23
## Setting seed: 125
## Applying Algorithm sann ...

## Warning: bounds can only be used with method L-BFGS-B (or Brent)

## Writing result file: /tmp/RtmpVeetNu/file576b3d599d5c/jobs/03/3-result.RData
## 2014-02-07 19:02:23: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:23: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/file576b3d599d5c/registry.RData
## Loading conf: /tmp/RtmpVeetNu/file576b3d599d5c/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=4 ##########
## Timestamp: 2014-02-07 19:02:23
## Setting seed: 126
## Applying Algorithm sann ...

## Warning: bounds can only be used with method L-BFGS-B (or Brent)

## Writing result file: /tmp/RtmpVeetNu/file576b3d599d5c/jobs/04/4-result.RData
## 2014-02-07 19:02:23: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:23: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/file576b3d599d5c/registry.RData
## Loading conf: /tmp/RtmpVeetNu/file576b3d599d5c/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=5 ##########
## Timestamp: 2014-02-07 19:02:23
## Setting seed: 127
## Applying Algorithm sann ...

## Warning: bounds can only be used with method L-BFGS-B (or Brent)

## Writing result file: /tmp/RtmpVeetNu/file576b3d599d5c/jobs/05/5-result.RData
## 2014-02-07 19:02:24: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:24: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/file576b3d599d5c/registry.RData
## Loading conf: /tmp/RtmpVeetNu/file576b3d599d5c/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=6 ##########
## Timestamp: 2014-02-07 19:02:24
## Setting seed: 128
## Applying Algorithm sann ...

## Warning: bounds can only be used with method L-BFGS-B (or Brent)

## Writing result file: /tmp/RtmpVeetNu/file576b3d599d5c/jobs/06/6-result.RData
## 2014-02-07 19:02:24: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:24: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/file576b3d599d5c/registry.RData
## Loading conf: /tmp/RtmpVeetNu/file576b3d599d5c/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=7 ##########
## Timestamp: 2014-02-07 19:02:24
## Setting seed: 129
## Applying Algorithm sann ...

## Warning: bounds can only be used with method L-BFGS-B (or Brent)

## Writing result file: /tmp/RtmpVeetNu/file576b3d599d5c/jobs/07/7-result.RData
## 2014-02-07 19:02:24: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:24: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/file576b3d599d5c/registry.RData
## Loading conf: /tmp/RtmpVeetNu/file576b3d599d5c/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=8 ##########
## Timestamp: 2014-02-07 19:02:24
## Setting seed: 130
## Applying Algorithm sann ...

## Warning: bounds can only be used with method L-BFGS-B (or Brent)

## Writing result file: /tmp/RtmpVeetNu/file576b3d599d5c/jobs/08/8-result.RData
## 2014-02-07 19:02:24: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:24: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/file576b3d599d5c/registry.RData
## Loading conf: /tmp/RtmpVeetNu/file576b3d599d5c/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=9 ##########
## Timestamp: 2014-02-07 19:02:24
## Setting seed: 131
## Applying Algorithm sann ...

## Warning: bounds can only be used with method L-BFGS-B (or Brent)

## Writing result file: /tmp/RtmpVeetNu/file576b3d599d5c/jobs/09/9-result.RData
## 2014-02-07 19:02:24: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:24: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/file576b3d599d5c/registry.RData
## Loading conf: /tmp/RtmpVeetNu/file576b3d599d5c/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=10 ##########
## Timestamp: 2014-02-07 19:02:24
## Setting seed: 132
## Applying Algorithm sann ...

## Warning: bounds can only be used with method L-BFGS-B (or Brent)

## Writing result file: /tmp/RtmpVeetNu/file576b3d599d5c/jobs/10/10-result.RData
## 2014-02-07 19:02:25: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:25: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/file576b3d599d5c/registry.RData
## Loading conf: /tmp/RtmpVeetNu/file576b3d599d5c/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=11 ##########
## Timestamp: 2014-02-07 19:02:25
## Setting seed: 133
## Applying Algorithm sann ...

## Warning: bounds can only be used with method L-BFGS-B (or Brent)

## Writing result file: /tmp/RtmpVeetNu/file576b3d599d5c/jobs/11/11-result.RData
## 2014-02-07 19:02:25: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:25: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/file576b3d599d5c/registry.RData
## Loading conf: /tmp/RtmpVeetNu/file576b3d599d5c/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=12 ##########
## Timestamp: 2014-02-07 19:02:25
## Setting seed: 134
## Applying Algorithm sann ...

## Warning: bounds can only be used with method L-BFGS-B (or Brent)

## Writing result file: /tmp/RtmpVeetNu/file576b3d599d5c/jobs/12/12-result.RData
## 2014-02-07 19:02:25: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:25: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/file576b3d599d5c/registry.RData
## Loading conf: /tmp/RtmpVeetNu/file576b3d599d5c/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=13 ##########
## Timestamp: 2014-02-07 19:02:25
## Setting seed: 135
## Applying Algorithm sann ...

## Warning: bounds can only be used with method L-BFGS-B (or Brent)

## Writing result file: /tmp/RtmpVeetNu/file576b3d599d5c/jobs/13/13-result.RData
## 2014-02-07 19:02:25: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:25: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/file576b3d599d5c/registry.RData
## Loading conf: /tmp/RtmpVeetNu/file576b3d599d5c/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=14 ##########
## Timestamp: 2014-02-07 19:02:25
## Setting seed: 136
## Applying Algorithm sann ...

## Warning: bounds can only be used with method L-BFGS-B (or Brent)

## Writing result file: /tmp/RtmpVeetNu/file576b3d599d5c/jobs/14/14-result.RData
## 2014-02-07 19:02:25: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:25: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/file576b3d599d5c/registry.RData
## Loading conf: /tmp/RtmpVeetNu/file576b3d599d5c/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=15 ##########
## Timestamp: 2014-02-07 19:02:25
## Setting seed: 137
## Applying Algorithm sann ...

## Warning: bounds can only be used with method L-BFGS-B (or Brent)

## Writing result file: /tmp/RtmpVeetNu/file576b3d599d5c/jobs/15/15-result.RData
## 2014-02-07 19:02:26: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:26: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/file576b3d599d5c/registry.RData
## Loading conf: /tmp/RtmpVeetNu/file576b3d599d5c/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=16 ##########
## Timestamp: 2014-02-07 19:02:26
## Setting seed: 138
## Applying Algorithm sann ...

## Warning: bounds can only be used with method L-BFGS-B (or Brent)

## Writing result file: /tmp/RtmpVeetNu/file576b3d599d5c/jobs/16/16-result.RData
## 2014-02-07 19:02:26: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:26: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/file576b3d599d5c/registry.RData
## Loading conf: /tmp/RtmpVeetNu/file576b3d599d5c/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=17 ##########
## Timestamp: 2014-02-07 19:02:26
## Setting seed: 139
## Applying Algorithm sann ...

## Warning: bounds can only be used with method L-BFGS-B (or Brent)

## Writing result file: /tmp/RtmpVeetNu/file576b3d599d5c/jobs/17/17-result.RData
## 2014-02-07 19:02:26: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:26: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/file576b3d599d5c/registry.RData
## Loading conf: /tmp/RtmpVeetNu/file576b3d599d5c/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=18 ##########
## Timestamp: 2014-02-07 19:02:26
## Setting seed: 140
## Applying Algorithm sann ...

## Warning: bounds can only be used with method L-BFGS-B (or Brent)

## Writing result file: /tmp/RtmpVeetNu/file576b3d599d5c/jobs/18/18-result.RData
## 2014-02-07 19:02:26: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:26: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/file576b3d599d5c/registry.RData
## Loading conf: /tmp/RtmpVeetNu/file576b3d599d5c/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=19 ##########
## Timestamp: 2014-02-07 19:02:26
## Setting seed: 141
## Applying Algorithm sann ...

## Warning: bounds can only be used with method L-BFGS-B (or Brent)

## Writing result file: /tmp/RtmpVeetNu/file576b3d599d5c/jobs/19/19-result.RData
## 2014-02-07 19:02:26: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:26: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/file576b3d599d5c/registry.RData
## Loading conf: /tmp/RtmpVeetNu/file576b3d599d5c/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=20 ##########
## Timestamp: 2014-02-07 19:02:26
## Setting seed: 142
## Applying Algorithm sann ...

## Warning: bounds can only be used with method L-BFGS-B (or Brent)

## Writing result file: /tmp/RtmpVeetNu/file576b3d599d5c/jobs/20/20-result.RData
## 2014-02-07 19:02:27: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:

## 
SubmitJobs |+                                                |   0% (00:00:00)
SubmitJobs |++                                               |   5% (00:00:00)
SubmitJobs |+++++                                            |  10% (00:00:00)
SubmitJobs |+++++++                                          |  15% (00:00:00)
SubmitJobs |++++++++++                                       |  20% (00:00:00)
SubmitJobs |++++++++++++                                     |  25% (00:00:03)
SubmitJobs |+++++++++++++++                                  |  30% (00:00:02)
SubmitJobs |+++++++++++++++++                                |  35% (00:00:01)
SubmitJobs |++++++++++++++++++++                             |  40% (00:00:01)
SubmitJobs |++++++++++++++++++++++                           |  45% (00:00:01)
SubmitJobs |++++++++++++++++++++++++                         |  50% (00:00:02)
SubmitJobs |+++++++++++++++++++++++++++                      |  55% (00:00:01)
SubmitJobs |+++++++++++++++++++++++++++++                    |  60% (00:00:01)
SubmitJobs |++++++++++++++++++++++++++++++++                 |  65% (00:00:01)
SubmitJobs |++++++++++++++++++++++++++++++++++               |  70% (00:00:00)
SubmitJobs |+++++++++++++++++++++++++++++++++++++            |  75% (00:00:01)
SubmitJobs |+++++++++++++++++++++++++++++++++++++++          |  80% (00:00:00)
SubmitJobs |++++++++++++++++++++++++++++++++++++++++++       |  85% (00:00:00)
SubmitJobs |++++++++++++++++++++++++++++++++++++++++++++     |  90% (00:00:00)
SubmitJobs |+++++++++++++++++++++++++++++++++++++++++++++++  |  95% (00:00:00)
SubmitJobs |+++++++++++++++++++++++++++++++++++++++++++++++++| 100% (00:00:00)

## Sending 20 submit messages...
## Might take some time, do not interrupt this!

# Gather informations from the experiments, in this case function value
# and whether the algorithm convergenced:
reduceResultsExperiments(reg, fun=function(job, res) res[c("value", "convergence")])

## Reducing 20 results...

## 
reduceResultsExperiments |+                                  |   0% (00:00:00)
reduceResultsExperiments |+++++++++++++++++++++++++++++++++++| 100% (00:00:00)

##    id     prob algo repl      value convergence
## 1   1 testfun1 sann    1  1.582e-40           0
## 2   2 testfun1 sann    2  1.582e-40           0
## 3   3 testfun1 sann    3  1.582e-40           0
## 4   4 testfun1 sann    4  1.582e-40           0
## 5   5 testfun1 sann    5  1.582e-40           0
## 6   6 testfun1 sann    6  1.582e-40           0
## 7   7 testfun1 sann    7  1.582e-40           0
## 8   8 testfun1 sann    8  1.582e-40           0
## 9   9 testfun1 sann    9  1.582e-40           0
## 10 10 testfun1 sann   10  1.582e-40           0
## 11 11 testfun2 sann    1 -2.061e-09           0
## 12 12 testfun2 sann    2 -2.061e-09           0
## 13 13 testfun2 sann    3 -2.061e-09           0
## 14 14 testfun2 sann    4 -2.061e-09           0
## 15 15 testfun2 sann    5 -2.061e-09           0
## 16 16 testfun2 sann    6 -2.061e-09           0
## 17 17 testfun2 sann    7 -2.061e-09           0
## 18 18 testfun2 sann    8 -2.061e-09           0
## 19 19 testfun2 sann    9 -2.061e-09           0
## 20 20 testfun2 sann   10 -2.061e-09           0

[Package BatchExperiments version 1.1 Index]