findExperiments {BatchExperiments}	R Documentation

Find ids of experiments that match a query.

Description

Find ids of experiments that match a query.

Usage

findExperiments(reg, ids, prob.pattern, prob.pars, algo.pattern, algo.pars,
  repls, match.substring = TRUE)

Arguments

reg
[ExperimentRegistry]
Registry.

ids
[integer]
Ids of selected experiments to restrict to. Default is all experiments.

prob.pattern
[character(1)]
If not missing, all problem ids that match this string are selected.

prob.pars
[R expression]
If not missing, all problems whose parameters match the given expression are selected.

algo.pattern
[character(1)]
If not missing, all algorithm ids that match this string are selected.

algo.pars
[R expression]
If not missing, all algorithms whose parameters match the given expression are selected.

repls
[integer]
If not missing, restrict to jobs with given replication numbers.

match.substring
[logical(1)]
Is a match in prob.pattern and algo.pattern if the id contains the pattern as substring or must the id exactly match? Default is TRUE.

Value

[integer]. Ids for experiments which match the query.

Examples

reg = makeExperimentRegistry(id="example1", file.dir=tempfile())

## Creating dir: /tmp/RtmpVeetNu/file576b452d0fbc
## Saving registry: /tmp/RtmpVeetNu/file576b452d0fbc/registry.RData

p1 = addProblem(reg, "one", 1)

## Writing problem files: /tmp/RtmpVeetNu/file576b452d0fbc/problems/one_static.RData, /tmp/RtmpVeetNu/file576b452d0fbc/problems/one_dynamic.RData

p2 = addProblem(reg, "two", 2)

## Writing problem files: /tmp/RtmpVeetNu/file576b452d0fbc/problems/two_static.RData, /tmp/RtmpVeetNu/file576b452d0fbc/problems/two_dynamic.RData

a = addAlgorithm(reg, "A", fun=function(static, n) static + n)

## Writing algorithm file: /tmp/RtmpVeetNu/file576b452d0fbc/algorithms/@A.RData

addExperiments(reg, algo.design=makeDesign(a, exhaustive = list(n = 1:4)))

## Adding 8 experiments / 8 jobs to DB.

findExperiments(reg, prob.pattern="one")

## [1] 1 2 3 4

findExperiments(reg, prob.pattern="o")

## [1] 1 2 3 4 5 6 7 8

findExperiments(reg, algo.pars = (n > 2))

## [1] 3 4 7 8

[Package BatchExperiments version 1.1 Index]