getIndex {BatchExperiments}	R Documentation

Group experiments.

Description

Creates a list of factor to use in functions like tapply, by or aggregate.

Usage

getIndex(reg, ids, by.prob = FALSE, by.algo = FALSE, by.repl = FALSE,
  by.prob.pars, by.algo.pars, enclos = parent.frame())

Arguments

reg
[ExperimentRegistry]
Registry.

ids
[integer]
If not missing, restict grouping to this subset of experiment ids.

by.prob
[logical]
Group experiments by problem. Default is FALSE.

by.algo
[logical]
Group experiments by algorithm. Default is FALSE.

by.repl
[logical]
Group experiments by replication. Default is FALSE.

by.prob.pars
[R expression]
If not missing, group experiments by this R expression. The expression is evaluated in the environment of problem parameters and converted to a factor using as.factor.

by.algo.pars
[R expression]
If not missing, group experiments by this R expression. The expression is evaluated in the environment of algorithm parameters and converted to a factor using as.factor.

enclos
[environment]
Enclosing frame for evaluation of parameters used by by.prob.pars and by.algo.pars, see eval. Defaults to the parent frame.

Value

[list]. List of factors.

Examples

# create a registry and add problems and algorithms
reg = makeExperimentRegistry("getIndex", file.dir=tempfile(""))

## Creating dir: /tmp/RtmpVeetNu/576b1ad2c155
## Saving registry: /tmp/RtmpVeetNu/576b1ad2c155/registry.RData

addProblem(reg, "prob", static = 1)

## Writing problem files: /tmp/RtmpVeetNu/576b1ad2c155/problems/prob_static.RData, /tmp/RtmpVeetNu/576b1ad2c155/problems/prob_dynamic.RData

addAlgorithm(reg, "f0", function(static, dynamic) static)

## Writing algorithm file: /tmp/RtmpVeetNu/576b1ad2c155/algorithms/f0.RData

addAlgorithm(reg, "f1", function(static, dynamic, i, k) static * i^k)

## Writing algorithm file: /tmp/RtmpVeetNu/576b1ad2c155/algorithms/f1.RData

ad = list(makeDesign("f0"), makeDesign("f1", exhaustive=list(i=1:10, k=1:3)))
addExperiments(reg, algo.designs=ad)

## Adding 31 experiments / 31 jobs to DB.

submitJobs(reg)

## Saving conf: /tmp/RtmpVeetNu/576b1ad2c155/conf.RData
## Submitting 31 chunks / 31 jobs.
## Cluster functions: Interactive.
## Auto-mailer settings: start=none, done=none, error=none.
## Writing 31 R scripts...
## 2014-02-07 19:02:28: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/576b1ad2c155/registry.RData
## Loading conf: /tmp/RtmpVeetNu/576b1ad2c155/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=1 ##########
## Timestamp: 2014-02-07 19:02:28
## Setting seed: 938655723
## Applying Algorithm f0 ...
## Writing result file: /tmp/RtmpVeetNu/576b1ad2c155/jobs/01/1-result.RData
## 2014-02-07 19:02:28: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:28: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/576b1ad2c155/registry.RData
## Loading conf: /tmp/RtmpVeetNu/576b1ad2c155/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=2 ##########
## Timestamp: 2014-02-07 19:02:28
## Setting seed: 938655724
## Applying Algorithm f1 ...
## Writing result file: /tmp/RtmpVeetNu/576b1ad2c155/jobs/02/2-result.RData
## 2014-02-07 19:02:28: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:29: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/576b1ad2c155/registry.RData
## Loading conf: /tmp/RtmpVeetNu/576b1ad2c155/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=3 ##########
## Timestamp: 2014-02-07 19:02:29
## Setting seed: 938655725
## Applying Algorithm f1 ...
## Writing result file: /tmp/RtmpVeetNu/576b1ad2c155/jobs/03/3-result.RData
## 2014-02-07 19:02:29: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:29: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/576b1ad2c155/registry.RData
## Loading conf: /tmp/RtmpVeetNu/576b1ad2c155/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=4 ##########
## Timestamp: 2014-02-07 19:02:29
## Setting seed: 938655726
## Applying Algorithm f1 ...
## Writing result file: /tmp/RtmpVeetNu/576b1ad2c155/jobs/04/4-result.RData
## 2014-02-07 19:02:29: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:29: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/576b1ad2c155/registry.RData
## Loading conf: /tmp/RtmpVeetNu/576b1ad2c155/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=5 ##########
## Timestamp: 2014-02-07 19:02:29
## Setting seed: 938655727
## Applying Algorithm f1 ...
## Writing result file: /tmp/RtmpVeetNu/576b1ad2c155/jobs/05/5-result.RData
## 2014-02-07 19:02:29: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:29: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/576b1ad2c155/registry.RData
## Loading conf: /tmp/RtmpVeetNu/576b1ad2c155/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=6 ##########
## Timestamp: 2014-02-07 19:02:29
## Setting seed: 938655728
## Applying Algorithm f1 ...
## Writing result file: /tmp/RtmpVeetNu/576b1ad2c155/jobs/06/6-result.RData
## 2014-02-07 19:02:29: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:29: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/576b1ad2c155/registry.RData
## Loading conf: /tmp/RtmpVeetNu/576b1ad2c155/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=7 ##########
## Timestamp: 2014-02-07 19:02:29
## Setting seed: 938655729
## Applying Algorithm f1 ...
## Writing result file: /tmp/RtmpVeetNu/576b1ad2c155/jobs/07/7-result.RData
## 2014-02-07 19:02:29: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:29: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/576b1ad2c155/registry.RData
## Loading conf: /tmp/RtmpVeetNu/576b1ad2c155/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=8 ##########
## Timestamp: 2014-02-07 19:02:29
## Setting seed: 938655730
## Applying Algorithm f1 ...
## Writing result file: /tmp/RtmpVeetNu/576b1ad2c155/jobs/08/8-result.RData
## 2014-02-07 19:02:29: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:30: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/576b1ad2c155/registry.RData
## Loading conf: /tmp/RtmpVeetNu/576b1ad2c155/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=9 ##########
## Timestamp: 2014-02-07 19:02:30
## Setting seed: 938655731
## Applying Algorithm f1 ...
## Writing result file: /tmp/RtmpVeetNu/576b1ad2c155/jobs/09/9-result.RData
## 2014-02-07 19:02:30: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:30: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/576b1ad2c155/registry.RData
## Loading conf: /tmp/RtmpVeetNu/576b1ad2c155/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=10 ##########
## Timestamp: 2014-02-07 19:02:30
## Setting seed: 938655732
## Applying Algorithm f1 ...
## Writing result file: /tmp/RtmpVeetNu/576b1ad2c155/jobs/10/10-result.RData
## 2014-02-07 19:02:30: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:30: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/576b1ad2c155/registry.RData
## Loading conf: /tmp/RtmpVeetNu/576b1ad2c155/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=11 ##########
## Timestamp: 2014-02-07 19:02:30
## Setting seed: 938655733
## Applying Algorithm f1 ...
## Writing result file: /tmp/RtmpVeetNu/576b1ad2c155/jobs/11/11-result.RData
## 2014-02-07 19:02:30: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:30: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/576b1ad2c155/registry.RData
## Loading conf: /tmp/RtmpVeetNu/576b1ad2c155/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=12 ##########
## Timestamp: 2014-02-07 19:02:30
## Setting seed: 938655734
## Applying Algorithm f1 ...
## Writing result file: /tmp/RtmpVeetNu/576b1ad2c155/jobs/12/12-result.RData
## 2014-02-07 19:02:30: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:30: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/576b1ad2c155/registry.RData
## Loading conf: /tmp/RtmpVeetNu/576b1ad2c155/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=13 ##########
## Timestamp: 2014-02-07 19:02:30
## Setting seed: 938655735
## Applying Algorithm f1 ...
## Writing result file: /tmp/RtmpVeetNu/576b1ad2c155/jobs/13/13-result.RData
## 2014-02-07 19:02:30: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:30: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/576b1ad2c155/registry.RData
## Loading conf: /tmp/RtmpVeetNu/576b1ad2c155/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=14 ##########
## Timestamp: 2014-02-07 19:02:30
## Setting seed: 938655736
## Applying Algorithm f1 ...
## Writing result file: /tmp/RtmpVeetNu/576b1ad2c155/jobs/14/14-result.RData
## 2014-02-07 19:02:30: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:31: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/576b1ad2c155/registry.RData
## Loading conf: /tmp/RtmpVeetNu/576b1ad2c155/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=15 ##########
## Timestamp: 2014-02-07 19:02:31
## Setting seed: 938655737
## Applying Algorithm f1 ...
## Writing result file: /tmp/RtmpVeetNu/576b1ad2c155/jobs/15/15-result.RData
## 2014-02-07 19:02:31: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:31: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/576b1ad2c155/registry.RData
## Loading conf: /tmp/RtmpVeetNu/576b1ad2c155/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=16 ##########
## Timestamp: 2014-02-07 19:02:31
## Setting seed: 938655738
## Applying Algorithm f1 ...
## Writing result file: /tmp/RtmpVeetNu/576b1ad2c155/jobs/16/16-result.RData
## 2014-02-07 19:02:31: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:31: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/576b1ad2c155/registry.RData
## Loading conf: /tmp/RtmpVeetNu/576b1ad2c155/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=17 ##########
## Timestamp: 2014-02-07 19:02:31
## Setting seed: 938655739
## Applying Algorithm f1 ...
## Writing result file: /tmp/RtmpVeetNu/576b1ad2c155/jobs/17/17-result.RData
## 2014-02-07 19:02:31: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:31: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/576b1ad2c155/registry.RData
## Loading conf: /tmp/RtmpVeetNu/576b1ad2c155/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=18 ##########
## Timestamp: 2014-02-07 19:02:31
## Setting seed: 938655740
## Applying Algorithm f1 ...
## Writing result file: /tmp/RtmpVeetNu/576b1ad2c155/jobs/18/18-result.RData
## 2014-02-07 19:02:31: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:31: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/576b1ad2c155/registry.RData
## Loading conf: /tmp/RtmpVeetNu/576b1ad2c155/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=19 ##########
## Timestamp: 2014-02-07 19:02:31
## Setting seed: 938655741
## Applying Algorithm f1 ...
## Writing result file: /tmp/RtmpVeetNu/576b1ad2c155/jobs/19/19-result.RData
## 2014-02-07 19:02:31: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:31: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/576b1ad2c155/registry.RData
## Loading conf: /tmp/RtmpVeetNu/576b1ad2c155/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=20 ##########
## Timestamp: 2014-02-07 19:02:31
## Setting seed: 938655742
## Applying Algorithm f1 ...
## Writing result file: /tmp/RtmpVeetNu/576b1ad2c155/jobs/20/20-result.RData
## 2014-02-07 19:02:31: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:32: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/576b1ad2c155/registry.RData
## Loading conf: /tmp/RtmpVeetNu/576b1ad2c155/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=21 ##########
## Timestamp: 2014-02-07 19:02:32
## Setting seed: 938655743
## Applying Algorithm f1 ...
## Writing result file: /tmp/RtmpVeetNu/576b1ad2c155/jobs/21/21-result.RData
## 2014-02-07 19:02:32: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:32: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/576b1ad2c155/registry.RData
## Loading conf: /tmp/RtmpVeetNu/576b1ad2c155/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=22 ##########
## Timestamp: 2014-02-07 19:02:32
## Setting seed: 938655744
## Applying Algorithm f1 ...
## Writing result file: /tmp/RtmpVeetNu/576b1ad2c155/jobs/22/22-result.RData
## 2014-02-07 19:02:32: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:32: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/576b1ad2c155/registry.RData
## Loading conf: /tmp/RtmpVeetNu/576b1ad2c155/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=23 ##########
## Timestamp: 2014-02-07 19:02:32
## Setting seed: 938655745
## Applying Algorithm f1 ...
## Writing result file: /tmp/RtmpVeetNu/576b1ad2c155/jobs/23/23-result.RData
## 2014-02-07 19:02:32: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:32: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/576b1ad2c155/registry.RData
## Loading conf: /tmp/RtmpVeetNu/576b1ad2c155/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=24 ##########
## Timestamp: 2014-02-07 19:02:32
## Setting seed: 938655746
## Applying Algorithm f1 ...
## Writing result file: /tmp/RtmpVeetNu/576b1ad2c155/jobs/24/24-result.RData
## 2014-02-07 19:02:32: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:32: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/576b1ad2c155/registry.RData
## Loading conf: /tmp/RtmpVeetNu/576b1ad2c155/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=25 ##########
## Timestamp: 2014-02-07 19:02:32
## Setting seed: 938655747
## Applying Algorithm f1 ...
## Writing result file: /tmp/RtmpVeetNu/576b1ad2c155/jobs/25/25-result.RData
## 2014-02-07 19:02:32: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:32: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/576b1ad2c155/registry.RData
## Loading conf: /tmp/RtmpVeetNu/576b1ad2c155/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=26 ##########
## Timestamp: 2014-02-07 19:02:32
## Setting seed: 938655748
## Applying Algorithm f1 ...
## Writing result file: /tmp/RtmpVeetNu/576b1ad2c155/jobs/26/26-result.RData
## 2014-02-07 19:02:33: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:33: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/576b1ad2c155/registry.RData
## Loading conf: /tmp/RtmpVeetNu/576b1ad2c155/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=27 ##########
## Timestamp: 2014-02-07 19:02:33
## Setting seed: 938655749
## Applying Algorithm f1 ...
## Writing result file: /tmp/RtmpVeetNu/576b1ad2c155/jobs/27/27-result.RData
## 2014-02-07 19:02:33: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:33: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/576b1ad2c155/registry.RData
## Loading conf: /tmp/RtmpVeetNu/576b1ad2c155/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=28 ##########
## Timestamp: 2014-02-07 19:02:33
## Setting seed: 938655750
## Applying Algorithm f1 ...
## Writing result file: /tmp/RtmpVeetNu/576b1ad2c155/jobs/28/28-result.RData
## 2014-02-07 19:02:33: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:33: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/576b1ad2c155/registry.RData
## Loading conf: /tmp/RtmpVeetNu/576b1ad2c155/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=29 ##########
## Timestamp: 2014-02-07 19:02:33
## Setting seed: 938655751
## Applying Algorithm f1 ...
## Writing result file: /tmp/RtmpVeetNu/576b1ad2c155/jobs/29/29-result.RData
## 2014-02-07 19:02:33: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:33: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/576b1ad2c155/registry.RData
## Loading conf: /tmp/RtmpVeetNu/576b1ad2c155/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=30 ##########
## Timestamp: 2014-02-07 19:02:33
## Setting seed: 938655752
## Applying Algorithm f1 ...
## Writing result file: /tmp/RtmpVeetNu/576b1ad2c155/jobs/30/30-result.RData
## 2014-02-07 19:02:33: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:
## 2014-02-07 19:02:33: Starting job on node bischl-x1.
## Loading registry: /tmp/RtmpVeetNu/576b1ad2c155/registry.RData
## Loading conf: /tmp/RtmpVeetNu/576b1ad2c155/conf.RData
## Auto-mailer settings: start=none, done=none, error=none.
## Setting work dir: /tmp/pkgdocs
## ########## Executing jid=31 ##########
## Timestamp: 2014-02-07 19:02:33
## Setting seed: 938655753
## Applying Algorithm f1 ...
## Writing result file: /tmp/RtmpVeetNu/576b1ad2c155/jobs/31/31-result.RData
## 2014-02-07 19:02:33: Job on slave is finished.
## Setting work back to: /tmp/pkgdocs
## Memory usage according to gc:

## 
SubmitJobs |+                                                |   0% (00:00:00)
SubmitJobs |++                                               |   3% (00:00:00)
SubmitJobs |+++                                              |   6% (00:00:14)
SubmitJobs |+++++                                            |  10% (00:00:09)
SubmitJobs |++++++                                           |  13% (00:00:06)
SubmitJobs |++++++++                                         |  16% (00:00:05)
SubmitJobs |+++++++++                                        |  19% (00:00:04)
SubmitJobs |+++++++++++                                      |  23% (00:00:03)
SubmitJobs |+++++++++++++                                    |  26% (00:00:05)
SubmitJobs |++++++++++++++                                   |  29% (00:00:04)
SubmitJobs |++++++++++++++++                                 |  32% (00:00:04)
SubmitJobs |+++++++++++++++++                                |  35% (00:00:03)
SubmitJobs |+++++++++++++++++++                              |  39% (00:00:03)
SubmitJobs |+++++++++++++++++++++                            |  42% (00:00:02)
SubmitJobs |++++++++++++++++++++++                           |  45% (00:00:03)
SubmitJobs |++++++++++++++++++++++++                         |  48% (00:00:03)
SubmitJobs |+++++++++++++++++++++++++                        |  52% (00:00:02)
SubmitJobs |+++++++++++++++++++++++++++                      |  55% (00:00:02)
SubmitJobs |++++++++++++++++++++++++++++                     |  58% (00:00:02)
SubmitJobs |++++++++++++++++++++++++++++++                   |  61% (00:00:01)
SubmitJobs |++++++++++++++++++++++++++++++++                 |  65% (00:00:02)
SubmitJobs |+++++++++++++++++++++++++++++++++                |  68% (00:00:01)
SubmitJobs |+++++++++++++++++++++++++++++++++++              |  71% (00:00:01)
SubmitJobs |++++++++++++++++++++++++++++++++++++             |  74% (00:00:01)
SubmitJobs |++++++++++++++++++++++++++++++++++++++           |  77% (00:00:01)
SubmitJobs |++++++++++++++++++++++++++++++++++++++++         |  81% (00:00:00)
SubmitJobs |+++++++++++++++++++++++++++++++++++++++++        |  84% (00:00:00)
SubmitJobs |+++++++++++++++++++++++++++++++++++++++++++      |  87% (00:00:00)
SubmitJobs |++++++++++++++++++++++++++++++++++++++++++++     |  90% (00:00:00)
SubmitJobs |++++++++++++++++++++++++++++++++++++++++++++++   |  94% (00:00:00)
SubmitJobs |+++++++++++++++++++++++++++++++++++++++++++++++  |  97% (00:00:00)
SubmitJobs |+++++++++++++++++++++++++++++++++++++++++++++++++| 100% (00:00:00)

## Sending 31 submit messages...
## Might take some time, do not interrupt this!

# get grouped job ids
ids = getJobIds(reg)
by(ids, getIndex(reg, by.prob=TRUE, by.algo=TRUE), identity)

## prob: prob
## algo: f0
## [1] 1
## -------------------------------------------------------- 
## prob: prob
## algo: f1
##  [1]  2  3  4  5  6  7  8  9 10 11 12 13 14 15 16 17 18 19 20 21 22 23 24
## [24] 25 26 27 28 29 30 31

ids.f1 = findExperiments(reg, algo.pattern="f1")
by(ids.f1, getIndex(reg, ids.f1, by.algo.pars=(k == 1)), identity)

## algo.pars: (k == 1): FALSE
##  [1] 12 13 14 15 16 17 18 19 20 21 22 23 24 25 26 27 28 29 30 31
## -------------------------------------------------------- 
## algo.pars: (k == 1): TRUE
##  [1]  2  3  4  5  6  7  8  9 10 11

# groupwise reduction
ids.f1 = findExperiments(reg, algo.pattern="f1")
f = function(aggr, job, res) aggr + res
by(ids.f1, getIndex(reg, ids.f1, by.algo.pars=k), reduceResults, reg=reg, fun=f)

## Reducing 10 results...

## 
reduceResults |+                                             |   0% (00:00:00)
reduceResults |+++++                                         |  10% (00:00:00)
reduceResults |+++++++++                                     |  20% (00:00:00)
reduceResults |++++++++++++++                                |  30% (00:00:00)
reduceResults |++++++++++++++++++                            |  40% (00:00:00)
reduceResults |+++++++++++++++++++++++                       |  50% (00:00:00)
reduceResults |++++++++++++++++++++++++++++                  |  60% (00:00:00)
reduceResults |++++++++++++++++++++++++++++++++              |  70% (00:00:00)
reduceResults |+++++++++++++++++++++++++++++++++++++         |  80% (00:00:00)
reduceResults |+++++++++++++++++++++++++++++++++++++++++     |  90% (00:00:00)
reduceResults |++++++++++++++++++++++++++++++++++++++++++++++| 100% (00:00:00)

## Reducing 10 results...

## 
reduceResults |+                                             |   0% (00:00:00)
reduceResults |+++++                                         |  10% (00:00:00)
reduceResults |+++++++++                                     |  20% (00:00:00)
reduceResults |++++++++++++++                                |  30% (00:00:00)
reduceResults |++++++++++++++++++                            |  40% (00:00:00)
reduceResults |+++++++++++++++++++++++                       |  50% (00:00:00)
reduceResults |++++++++++++++++++++++++++++                  |  60% (00:00:00)
reduceResults |++++++++++++++++++++++++++++++++              |  70% (00:00:00)
reduceResults |+++++++++++++++++++++++++++++++++++++         |  80% (00:00:00)
reduceResults |+++++++++++++++++++++++++++++++++++++++++     |  90% (00:00:00)
reduceResults |++++++++++++++++++++++++++++++++++++++++++++++| 100% (00:00:00)

## Reducing 10 results...

## 
reduceResults |+                                             |   0% (00:00:00)
reduceResults |+++++                                         |  10% (00:00:00)
reduceResults |+++++++++                                     |  20% (00:00:00)
reduceResults |++++++++++++++                                |  30% (00:00:00)
reduceResults |++++++++++++++++++                            |  40% (00:00:00)
reduceResults |+++++++++++++++++++++++                       |  50% (00:00:00)
reduceResults |++++++++++++++++++++++++++++                  |  60% (00:00:00)
reduceResults |++++++++++++++++++++++++++++++++              |  70% (00:00:00)
reduceResults |+++++++++++++++++++++++++++++++++++++         |  80% (00:00:00)
reduceResults |+++++++++++++++++++++++++++++++++++++++++     |  90% (00:00:00)
reduceResults |++++++++++++++++++++++++++++++++++++++++++++++| 100% (00:00:00)

## algo.pars: k: 1
## [1] 55
## -------------------------------------------------------- 
## algo.pars: k: 2
## [1] 385
## -------------------------------------------------------- 
## algo.pars: k: 3
## [1] 3025

by(ids.f1, getIndex(reg, ids.f1, by.algo.pars=i), reduceResults, reg=reg, fun=f)

## Reducing 3 results...

## 
reduceResults |+                                             |   0% (00:00:00)
reduceResults |+++++++++++++++                               |  33% (00:00:00)
reduceResults |+++++++++++++++++++++++++++++++               |  67% (00:00:00)
reduceResults |++++++++++++++++++++++++++++++++++++++++++++++| 100% (00:00:00)

## Reducing 3 results...

## 
reduceResults |+                                             |   0% (00:00:00)
reduceResults |+++++++++++++++                               |  33% (00:00:00)
reduceResults |+++++++++++++++++++++++++++++++               |  67% (00:00:00)
reduceResults |++++++++++++++++++++++++++++++++++++++++++++++| 100% (00:00:00)

## Reducing 3 results...

## 
reduceResults |+                                             |   0% (00:00:00)
reduceResults |+++++++++++++++                               |  33% (00:00:00)
reduceResults |+++++++++++++++++++++++++++++++               |  67% (00:00:00)
reduceResults |++++++++++++++++++++++++++++++++++++++++++++++| 100% (00:00:00)

## Reducing 3 results...

## 
reduceResults |+                                             |   0% (00:00:00)
reduceResults |+++++++++++++++                               |  33% (00:00:00)
reduceResults |+++++++++++++++++++++++++++++++               |  67% (00:00:00)
reduceResults |++++++++++++++++++++++++++++++++++++++++++++++| 100% (00:00:00)

## Reducing 3 results...

## 
reduceResults |+                                             |   0% (00:00:00)
reduceResults |+++++++++++++++                               |  33% (00:00:00)
reduceResults |+++++++++++++++++++++++++++++++               |  67% (00:00:00)
reduceResults |++++++++++++++++++++++++++++++++++++++++++++++| 100% (00:00:00)

## Reducing 3 results...

## 
reduceResults |+                                             |   0% (00:00:00)
reduceResults |+++++++++++++++                               |  33% (00:00:00)
reduceResults |+++++++++++++++++++++++++++++++               |  67% (00:00:00)
reduceResults |++++++++++++++++++++++++++++++++++++++++++++++| 100% (00:00:00)

## Reducing 3 results...

## 
reduceResults |+                                             |   0% (00:00:00)
reduceResults |+++++++++++++++                               |  33% (00:00:00)
reduceResults |+++++++++++++++++++++++++++++++               |  67% (00:00:00)
reduceResults |++++++++++++++++++++++++++++++++++++++++++++++| 100% (00:00:00)

## Reducing 3 results...

## 
reduceResults |+                                             |   0% (00:00:00)
reduceResults |+++++++++++++++                               |  33% (00:00:00)
reduceResults |+++++++++++++++++++++++++++++++               |  67% (00:00:00)
reduceResults |++++++++++++++++++++++++++++++++++++++++++++++| 100% (00:00:00)

## Reducing 3 results...

## 
reduceResults |+                                             |   0% (00:00:00)
reduceResults |+++++++++++++++                               |  33% (00:00:00)
reduceResults |+++++++++++++++++++++++++++++++               |  67% (00:00:00)
reduceResults |++++++++++++++++++++++++++++++++++++++++++++++| 100% (00:00:00)

## Reducing 3 results...

## 
reduceResults |+                                             |   0% (00:00:00)
reduceResults |+++++++++++++++                               |  33% (00:00:00)
reduceResults |+++++++++++++++++++++++++++++++               |  67% (00:00:00)
reduceResults |++++++++++++++++++++++++++++++++++++++++++++++| 100% (00:00:00)

## algo.pars: i: 1
## [1] 3
## -------------------------------------------------------- 
## algo.pars: i: 2
## [1] 14
## -------------------------------------------------------- 
## algo.pars: i: 3
## [1] 39
## -------------------------------------------------------- 
## algo.pars: i: 4
## [1] 84
## -------------------------------------------------------- 
## algo.pars: i: 5
## [1] 155
## -------------------------------------------------------- 
## algo.pars: i: 6
## [1] 258
## -------------------------------------------------------- 
## algo.pars: i: 7
## [1] 399
## -------------------------------------------------------- 
## algo.pars: i: 8
## [1] 584
## -------------------------------------------------------- 
## algo.pars: i: 9
## [1] 819
## -------------------------------------------------------- 
## algo.pars: i: 10
## [1] 1110

[Package BatchExperiments version 1.1 Index]