makeDesign {BatchExperiments}	R Documentation

Create parameter designs for problems and algorithms.

Description

Create a parameter design for either a problem or an algorithm that you can use in addExperiments. All parameters in design and exhaustive be “primitive” in the sense that either is.atomic is TRUE or is.factor is TRUE.

Usage

makeDesign(id, design = data.frame(), exhaustive = list())

Arguments

id
[character(1)]
Id of algorithm or problem.

design
[data.frame]
The design. Must have named columns corresponding to parameters. Default is an empty data.frame().

exhaustive
[list]
Named list of parameters settings which should be exhaustively tried. Names must correspond to parameters. Default is empty list.

Details

Be aware of R's default behaviour of converting strings into factors if you use the design parameter. See option stringsAsFactors in data.frame to turn this off.

Value

[Design].

Examples

## Not run: 
##D # simple design for algorithm "a1" with no parameters:
##D design <- makeDesign("a1")
##D 
##D # design for problem "p1" using predefined parameter combinations
##D design <- makeDesign("p1", design = data.frame(alpha = 0:1, beta = c(0.1, 0.2)))
##D 
##D # creating a list of designs for several algorithms at once, all using the same
##D # exhaustive grid of parameters
##D designs <- lapply(c("a1", "a2", "a3"), makeDesign,
##D                   exhaustive = list(alpha = 0:1, gamma = 1:10/10))
## End(Not run)

[Package BatchExperiments version 1.1 Index]