reduceResultsExperiments {BatchExperiments}	R Documentation

Reduce results into a data.frame with all relevant information.

Description

Generates a data.frame with one row per job id. The columns are: ids of problem and algorithm (named “prob” and “algo”), one column per parameter of problem or algorithm (named by the parameter name), the replication number (named “repl”) and all columns defined in the function to collect the values. Note that you cannot rely on the order of the columns. If a paramater does not have a setting for a certain job / experiment it is set to NA. Have a look at getResultVars if you want to use somethink like ddply on the results.

Usage

reduceResultsExperiments(reg, ids, part = NA_character_, fun, ...,
  strings.as.factors = default.stringsAsFactors(), block.size, impute.val)

Arguments

reg
[ExperimentRegistry]
Registry.

ids
[integer]
Ids of selected experiments. Default is all jobs for which results are available.

part
[character] Only useful for multiple result files, then defines which result file part(s) should be loaded. NA means all parts are loaded, which is the default.

fun
[function(job, res, ...)]
Function to collect values from job and result res object, the latter from stored result file. Must return an object which can be coerced to a data.frame (e.g. a list). Default is a function that simply returns res which may or may not work, depending on the type of res.

...
[any]
Additional arguments to fun.

strings.as.factors
[logical(1)] Should all character columns in result be converted to factors? Default is default.stringsAsFactors().

block.size
[integer(1)] Results will be fetched in blocks of this size. Default is max(100, 5 percent of ids).

impute.val
[named list]
If not missing, the value of impute.val is used as a replacement for the return value of function fun on missing results. An empty list is allowed.

Value

[data.frame]. Aggregated results, containing problem and algorithm paramaters and collected values.

[Package BatchExperiments version 1.1 Index]