reduceResultsExperimentsParallel {BatchExperiments}	R Documentation

Reduce very many results in parallel.

Description

Basically the same as reduceResultsExperiments but creates a few (hopefully short) jobs to reduce the results in parallel. The function internally calls batchMapQuick, does “busy-waiting” till all jobs are done and cleans all temporary files up. Useful when you have very results and reducing is slow.

Usage

reduceResultsExperimentsParallel(reg, ids, part = as.character(NA), fun, ...,
  njobs = 20L, strings.as.factors = default.stringsAsFactors())

Arguments

reg
[ExperimentRegistry]
Registry.

ids
[integer]
Ids of selected experiments. Default is all jobs for which results are available.

part
[character] Only useful for multiple result files, then defines which result file part(s) should be loaded. NA means all parts are loaded, which is the default.

fun
[function(job, res, ...)]
Function to collect values from job and result res object, the latter from stored result file. Must return an object which can be coerced to a data.frame (e.g. a list). Default is a function that simply returns res which may or may not work, depending on the type of res.

...
[any]
Additional arguments to fun.

njobs
[integer(1)] Number of parallel jobs to create. Default is 20.

strings.as.factors
[logical(1)] Should all character columns in result be converted to factors? Default is default.stringsAsFactors().

Value

[data.frame]. Aggregated results, containing problem and algorithm paramaters and collected values.

[Package BatchExperiments version 1.1 Index]