removeExperiments {BatchExperiments}	R Documentation

Remove jobs from registry.

Description

THIS DELETES ALL FILES REGARDING THE JOBS, INCLUDING RESULTS! If you really know what you are doing, you may set force to TRUE to omit sanity checks on running jobs.

Usage

removeExperiments(reg, ids, force = FALSE)

Arguments

reg
[ExperimentRegistry]
Registry.

ids
[integer]
Ids of jobs you want to remove. Default is none.

force
[logical(1)]
Also remove jobs which seem to be still running. Default is FALSE.

Value

Vector of type integer of removed job ids.

[Package BatchExperiments version 1.1 Index]