summarizeExperiments {BatchExperiments}	R Documentation

Summarize selected experiments.

Description

A data.frame is returned that contains summary information about the selected experiments. The data.frame is constructed by building the columns “prob, <prob.pars>, algo, <algo.pars>, repl”, rbind.fill is used to connect the rows, so if some parameters do not appear in some experiments, they will be set to NA. Now only the columns in show will be selected, how many of such experiments exist will be counted in a new column “.count”.

Usage

summarizeExperiments(reg, ids, show = c("prob", "algo"))

Arguments

reg
[ExperimentRegistry]
Registry.

ids
[integer]
Selected experiments. Default is all experiments.

show
[character]
Should detailed information for each single experiment be printed? Default is c("prob", "algo").

Value

[data.frame].

Examples

reg <- makeExperimentRegistry("summarizeExperiments", seed=123, file.dir=tempfile())

## Creating dir: /tmp/RtmpVeetNu/file576b45e5681d
## Saving registry: /tmp/RtmpVeetNu/file576b45e5681d/registry.RData

p1 <- addProblem(reg, "p1", static=1)

## Writing problem files: /tmp/RtmpVeetNu/file576b45e5681d/problems/p1_static.RData, /tmp/RtmpVeetNu/file576b45e5681d/problems/p1_dynamic.RData

a1 <- addAlgorithm(reg, id="a1", fun=function(static, dynamic, alpha, beta) 1)

## Writing algorithm file: /tmp/RtmpVeetNu/file576b45e5681d/algorithms/a1.RData

a2 <- addAlgorithm(reg, id="a2", fun=function(static, dynamic, alpha, gamma) 2)

## Writing algorithm file: /tmp/RtmpVeetNu/file576b45e5681d/algorithms/a2.RData

ad1 <- makeDesign(a1, exhaustive = list(alpha=1:2, beta=1:2))
ad2 <- makeDesign(a2, exhaustive = list(alpha=1:2, gamma=7:8))
addExperiments(reg, algo.designs=list(ad1, ad2), repls=2)

## Adding 8 experiments / 16 jobs to DB.

print(summarizeExperiments(reg))

##   prob algo .count
## 1   p1   a1      8
## 2   p1   a2      8

print(summarizeExperiments(reg, show=c("prob", "algo", "alpha", "gamma")))

##   prob algo alpha gamma V1
## 1   p1   a1     1    NA  4
## 2   p1   a1     2    NA  4
## 3   p1   a2     1     7  2
## 4   p1   a2     1     8  2
## 5   p1   a2     2     7  2
## 6   p1   a2     2     8  2

[Package BatchExperiments version 1.1 Index]